+++ sites/docs/src/lib/stores/package-manager.ts
@@ -1,6 +1,11 @@
 import { getContext, setContext } from "svelte";
 import { persisted } from "svelte-persisted-store";
-import type { Agent } from "package-manager-detector";
+import {
+	resolveCommand,
+	type Agent,
+	type Command,
+	type ResolvedCommand,
+} from "package-manager-detector";
 
 const PACKAGE_MANAGER = Symbol("packageManager");
 
@@ -17,4 +22,29 @@ export function getPackageManager(): ReturnType<typeof setPackageManager> {
 function createPackageManagerStore(key: string, initialValue: Agent) {
 	const store = persisted(key, initialValue);
 	return store;
+}
+
+export type PackageManagerCommand = Command | "create";
+
+export function getCommand(
+	pm: Agent,
+	type: PackageManagerCommand,
+	command: string | string[]
+): ResolvedCommand {
+	let args = [];
+	if (typeof command === "string") {
+		args = command.split(" ");
+	} else {
+		args = command;
+	}
+
+	// special handling for create
+	if (type === "create") return { command: pm, args: ["create", ...args] };
+
+	const cmd = resolveCommand(pm, type, args);
+
+	// since docs are static any unresolved command is a code error
+	if (cmd === null) throw new Error("Could not resolve command!");
+
+	return cmd;
 }
